Code to play hangman
Added:

Array with test words.

Get a random word

Let the user guess a letter

Calculate if the user did lose with the amount of fails ( per letter that the user not already guessed, and is not in the word ).

Show the word every turn to the user.

Show if the user won, or lost on the end.